added a good amount
Created the natural flow for the program. Ask the user for input
on relevence feedback, and then based on the precision that is calculated
we will branch to a particular control flow.

I just implemented the edge cases, not the actual algorithm (i.e. the whole
point of the assignment), which modifies the query to re-submit to google
to hopefully increase the relevence precision.

This program is in good shape right now.