Created a Parse Object that tokenizes html docs
This module will take in a url and perform a get request, it will
regex match, make all words lower, check for stopwords and numeric
words and return the list of words.

Now, If the get request fails then it will throw an exception
that is caught in google-query which will then use the snippet
and title data that is already stored.

So there is a built in error checking. We want to be able to use
the whole document, and that is the goal. The code is a little slow
now, idk if that is in response to my algo, or becuase we are simply
processing large html documents now.