added in parsing for the whole html document
code is a little slower, need to see where the bottleneck is.
I think it is simply cuz we are parsing large html files now.

Moved, stopword check code into the Tokenizer class.
Now use the Tokenizer class to parse the html file using a get
request, and if it fails, then use the data that we are already
storing on hand.

The code is definitely cleaner and easier to understand now.