double checked math
I double checked the math for tf-idf weighting and it is correct.
I also changed the way we go through the pandas dataframe because
certain operations are optimized and faster. Thus, I changed around
some of the functions to make the code more readable.

We are definitely ready for the next step now, and should try to come
up with something soon and then iterate on it...