generated new potential query lists
This uses all of the words pulled from step 1.

Now we have to test all of the new potential queries using rocchio's
algo, and return the best one.

I think this is close. Now that we can return a brand new query
after some ML work, we can refine the weights and test against
the prof executable.