added skip for bad sentences
for efficieny purposes, if a sentence cannot extract entities that
cover the subject and the object at a minimum then we will skip.

The check function, wants to know if we can satisfy a singular pair.
Of course, will not care about how many pairs, just that we have >= 1.

Now onto Bert ...