big commit
This code seems to be essentially done.

1) changed the way we seed query, use the string no need for splitting
into the keywords (makes for harder comparisons later on).

2) if at the end of going through the links in an iter we dont have
enough tuples pushed logic to figure out how to proceed.

	2.1) if we stall, then quit and return all the extracted tuples
	2.2) if we find a new tuple (not yet queried) lets use that as a
	seed for the next round.

3) added sort logic and displaying for the tuples to the console.

This code works, and will produce all of the relations that are necessary.
We will just need to double check some of the logic, and ensure that
we are hitting all of the edge cases.

Ethan, we seem to be pulling in a lot of tuples. We have to double
check that they are all correct. Need to go back into the Tokenizer code

when you want to test, just throw a quit() in there to step through
the code.